+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- scatter_ray method in BRDF: SpecularBRDF, DiffusiveBRDF
+- OrthoNormal Base generator
 - Added flag `--algorithm` to choose between OnOffRenderer or FlatRenderer
 - Renderer struct with OnOffRenderer and FlatRenderer
 - Material, BRDF, Pigments (Image, Checkered, Uniform) struct